+++ code/Input/movement.js
@@ -39,6 +39,9 @@ function movement() {
       }
     }
     dasFrameLeft ++;
+    if (areFrame > 0) {
+      dasFrameLeft = das;
+    }
     dasFrameRight = 0;
   } else if (axisX == 1) {
     if (dasFrameRight === 0 || dasFrameRight >= das) {
@@ -51,7 +54,9 @@ function movement() {
     dasFrameRight ++;
     dasFrameLeft = 0;
   } else {
-    dasFrameLeft = 0;
+    if (areFrame > 0) {
+      dasFrameRight = das;
+    }
     dasFrameRight = 0;
   }